Support display scales other than 100%/200% on Windows. IGraphics::mScreenScale becomes a float
Without this change the plug-in was not the correct size on displays where the scaling was e.g. 150%, it was quantized

+++ IGraphics/Platforms/IGraphicsWin.h
@@ -34,7 +34,7 @@ public:
   void* GetWinModuleHandle() override { return mHInstance; }
 
   void ForceEndUserEdit() override;
-  int GetPlatformWindowScale() const override { return GetScreenScale(); }
+  float GetPlatformWindowScale() const override { return GetScreenScale(); }
 
   void PlatformResize(bool parentHasResized) override;
